- fps Danh

+++ trunk/SolutionContraGame/ContraGame/CTimeSpanDx9.h
@@ -17,7 +17,7 @@ public:
 	
 	CTimeSpanDx9(UINT64 Ticks);
 	
-	float	getMilliseconds();
+	int	getMilliseconds();
 	
 	UINT64	getTicks();
